Testing: Get first object key util tested, documentated and validation added
