+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "pomodoros",
-	"version": "0.5.2",
+	"version": "0.5.3",
 	"private": true,
 	"scripts": {
 		"dev": "next dev -p 3002",
