+++ src/utils/getFirstObjectKey/index.ts
@@ -1,0 +1,3 @@
+import getFirstObjectKey from "./getFirstObjectKey";
+
+export default getFirstObjectKey;
